Add gateway module pattern and polish the todo list prototype
- Introduce build-logic (gateway.application convention plugin) and relocate
  app/androidApp to gateway/todoList, the first of several planned per-product
  Android app modules; each future gateway only needs its own identity block.
- Fix "Put on top" to actually prepend: priorityOrder is now Double, computed
  as (currentMin ?: 1.0) * 0.9 so new prioritized items always sort first
  without ever renumbering existing rows (previous Int scheme appended
  instead of prepending).
- Add safeDrawingPadding() to AppBackground so screen content and the FAB
  clear the status/nav bars app-wide.
- Add-todo sheet: sentence auto-capitalization, "Put on top"/"A todo text"
  copy, and defaulting to "Todo added at HH:mm" when submitted blank.

+++ app/shared/src/commonMain/kotlin/com/artemonre/onemoretodolist/feature/todolist/domain/TodoItem.kt
@@ -8,6 +8,8 @@ data class TodoItem(
     val isDone: Boolean,
     val sortOrder: Int,
     val date: LocalDate,
-    // Rank among other prioritized items; null means not prioritized. Only affects TodoSortOption.Date.
-    val priorityOrder: Int? = null
+    // Rank among other prioritized items; null means not prioritized, lower sorts first.
+    // New values are assigned as (currentMin ?: 1.0) * 0.9, always positive, never needing
+    // to renumber existing items. Only affects TodoSortOption.Date.
+    val priorityOrder: Double? = null
 )
